requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.3.0
pycodestyle==2.10.0
pymongo==4.3.3
python-dotenv==0.21.1
python-snappy==0.6.1
pytz==2022.7.1
sqlparse==0.2.4
tomli==2.0.1
tzdata==2022.7
# Uncomment if on windows
# winkerberos==0.9.1
